Using file on Aws instead of database
