+++ nogo/build.gradle
@@ -15,9 +15,7 @@ dependencies {
     compile 'org.springframework.boot:spring-boot-starter-actuator'
     compile 'org.springframework.boot:spring-boot-starter-web'
 
-    compile 'org.springframework.boot:spring-boot-starter-data-jpa'
     compile 'com.vividsolutions:jts:1.13'
-    compile 'mysql:mysql-connector-java:5.1.40'
 
     compile fileTree(dir: 'libs', include: '*.jar')
 
@@ -27,7 +25,6 @@ dependencies {
 
     compile "com.amazonaws:aws-java-sdk-s3:1.11.118"
 
-    testCompile 'com.h2database:h2:1.4.191'
     testCompile 'org.springframework.boot:spring-boot-starter-test'
 
 }
