--- nogo/src/main/java/dk/dma/nogoservice/repository/SydKattegatRepository.java
@@ -1,28 +1,0 @@
-/* Copyright (c) 2011 Danish Maritime Authority.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package dk.dma.nogoservice.repository;
-
-import dk.dma.nogoservice.entity.SouthKattegat;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-/**
- * @author Klaus Groenbaek
- *         Created 12/03/17.
- */
-@Repository
-public interface SydKattegatRepository extends JpaRepository<SouthKattegat, Integer> {
-
-}
